Move java-related code to the azure-intellij-plugin-keyvault-java library
