+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-keyvault-java/src/main/java/com/microsoft/azure/toolkit/intellij/keyvault/code/spring/EnvVarConfidence.java
