+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-keyvault-java/src/main/java/com/microsoft/azure/toolkit/intellij/keyvault/connection/KeyVaultResourceDefinition.java
@@ -6,15 +6,12 @@
 package com.microsoft.azure.toolkit.intellij.keyvault.connection;
 
 import com.intellij.openapi.project.Project;
-import com.microsoft.azure.toolkit.ide.common.icon.AzureIcons;
 import com.microsoft.azure.toolkit.intellij.common.AzureFormJPanel;
-import com.microsoft.azure.toolkit.intellij.connector.AzureServiceResource;
 import com.microsoft.azure.toolkit.intellij.connector.Connection;
 import com.microsoft.azure.toolkit.intellij.connector.Resource;
 import com.microsoft.azure.toolkit.intellij.connector.spring.SpringSupported;
 import com.microsoft.azure.toolkit.lib.Azure;
 import com.microsoft.azure.toolkit.lib.auth.AzureCloud;
-import com.microsoft.azure.toolkit.lib.keyvault.AzureKeyVault;
 import com.microsoft.azure.toolkit.lib.keyvault.KeyVault;
 import lombok.Getter;
 import org.apache.commons.lang3.StringUtils;
@@ -22,29 +19,18 @@ import org.apache.commons.lang3.tuple.Pair;
 
 import javax.annotation.Nullable;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
+
 
 @Getter
-public class KeyVaultResourceDefinition extends AzureServiceResource.Definition<KeyVault> implements SpringSupported<KeyVault> {
+public class KeyVaultResourceDefinition extends BaseKeyVaultResourceDefinition
+        implements SpringSupported<KeyVault> {
+
     public static final KeyVaultResourceDefinition INSTANCE = new KeyVaultResourceDefinition();
 
-    public KeyVaultResourceDefinition() {
-        super("Microsoft.KeyVault", "Azure Key Vault", AzureIcons.KeyVault.MODULE.getIconPath());
-    }
-
     @Override
-    public String getDefaultEnvPrefix() {
-        return "KEY_VAULT";
-    }
-
-    @Override
-    public Map<String, String> initEnv(AzureServiceResource<KeyVault> vaultDef, Project project) {
-        final KeyVault vault = vaultDef.getData();
-        final HashMap<String, String> env = new HashMap<>();
-        env.put(String.format("%s_ENDPOINT", Connection.ENV_PREFIX), vault.getVaultUri());
-        return env;
+    public AzureFormJPanel<Resource<KeyVault>> getResourcePanel(Project project) {
+        return new KeyVaultResourcePanel();
     }
 
     @Override
@@ -64,22 +50,5 @@ public class KeyVaultResourceDefinition extends AzureServiceResource.Definition<
             properties.add(Pair.of("spring.cloud.azure.keyvault.secret.property-source-enabled", "true"));
         }
         return properties;
-    }
-
-    @Override
-    public KeyVault getResource(String dataId, final String id) {
-        return Azure.az(AzureKeyVault.class).getById(dataId);
-    }
-
-    @Override
-    public List<Resource<KeyVault>> getResources(Project project) {
-        return Azure.az(AzureKeyVault.class).list().stream()
-            .flatMap(m -> m.getKeyVaultModule().list().stream())
-            .map(this::define).toList();
-    }
-
-    @Override
-    public AzureFormJPanel<Resource<KeyVault>> getResourcePanel(Project project) {
-        return new KeyVaultResourcePanel();
     }
 }
